+++ src/main/java/com/eduribeiro8/LilMarket/dto/CustomerPaymentResponseDTO.java
@@ -7,9 +7,9 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @Schema(description = "Informações do pagamento retornadas")
 public record CustomerPaymentResponseDTO(
         @Schema(description = "Identificador do pagamento", example = "12")
-        Integer paymentId,
+        Long paymentId,
         @Schema(description = "Identificador do cliente", example = "5")
-        Integer customerId,
+        Long customerId,
         @Schema(description = "Valor pago", example = "50.00")
         BigDecimal amountPaid,
         @Schema(description = "Método de pagamento", example = "DINHEIRO")
