+++ src/main/java/com/eduribeiro8/LilMarket/repository/SaleRepository.java
@@ -8,7 +8,7 @@ import java.time.OffsetDateTime;
 import java.util.List;
 
 @Repository
-public interface SaleRepository extends JpaRepository<Sale, Integer> {
+public interface SaleRepository extends JpaRepository<Sale, Long> {
 
     List<Sale> findByTimestampBetween(OffsetDateTime start, OffsetDateTime end);
 }
